Bug 758670 - Use stdint types instead of stdin ones in dom; r=mounir

+++ content/canvas/src/nsCanvasRenderingContext2D.cpp
@@ -4036,7 +4036,7 @@ nsCanvasRenderingContext2D::PutImageData_explicit(PRInt32 x, PRInt32 y, PRUint32
                 return NS_ERROR_DOM_INDEX_SIZE_ERR;
 
             dirtyX = checkedDirtyX.value();
-            dirtyWidth = -(int32)dirtyWidth;
+            dirtyWidth = -dirtyWidth;
         }
 
         if (dirtyHeight < 0) {
@@ -4048,7 +4048,7 @@ nsCanvasRenderingContext2D::PutImageData_explicit(PRInt32 x, PRInt32 y, PRUint32
                 return NS_ERROR_DOM_INDEX_SIZE_ERR;
 
             dirtyY = checkedDirtyY.value();
-            dirtyHeight = -(int32)dirtyHeight;
+            dirtyHeight = -dirtyHeight;
         }
 
         // bound the dirty rect within the imageData rectangle
